need to fix printf floats

+++ srcs/ft_put/ft_putwchar_fd.c
@@ -6,7 +6,7 @@
 /*   By: ttroll <ttroll@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/03 13:15:55 by ttroll            #+#    #+#             */
-/*   Updated: 2019/04/12 13:13:58 by ttroll           ###   ########.fr       */
+/*   Updated: 2019/04/22 18:59:05 by ttroll           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ int						ft_putwchar_fd(wint_t c, int fd)
 	size = (size_t)get_wchar_size(c);
 	if (size == 1)
 		i = ft_putchar_fd((char)c, fd);
-	else if (size > MB_CUR_MAX)
+	else if ((int)size > MB_CUR_MAX)
 		return (err_handle(c, fd));
 	else
 	{
